+++ api/applyconfiguration/api/v1alpha1/pod.go
@@ -17,7 +17,7 @@ type PodApplyConfiguration struct {
 	Affinity                      *v1.Affinity                            `json:"affinity,omitempty"`
 	Tolerations                   []v1.Toleration                         `json:"tolerations,omitempty"`
 	GracefulShutdown              *GracefulShutdownSpecApplyConfiguration `json:"gracefulShutdown,omitempty"`
-	TerminationGracePeriodSeconds *int                                    `json:"terminationGracePeriodSeconds,omitempty"`
+	TerminationGracePeriodSeconds *int64                                  `json:"terminationGracePeriodSeconds,omitempty"`
 	ReadinessProbe                *v1.Probe                               `json:"readinessProbe,omitempty"`
 	LivenessProbe                 *v1.Probe                               `json:"livenessProbe,omitempty"`
 	TopologySpreadConstraints     []v1.TopologySpreadConstraint           `json:"topologySpreadConstraints,omitempty"`
@@ -119,7 +119,7 @@ func (b *PodApplyConfiguration) WithGracefulShutdown(value *GracefulShutdownSpec
 // WithTerminationGracePeriodSeconds sets the TerminationGracePeriodSeconds field in the declarative configuration to the given value
 // and returns the receiver, so that objects can be built by chaining "With" function invocations.
 // If called multiple times, the TerminationGracePeriodSeconds field is set to the value of the last call.
-func (b *PodApplyConfiguration) WithTerminationGracePeriodSeconds(value int) *PodApplyConfiguration {
+func (b *PodApplyConfiguration) WithTerminationGracePeriodSeconds(value int64) *PodApplyConfiguration {
 	b.TerminationGracePeriodSeconds = &value
 	return b
 }
