+++ internal/kgateway/extensions2/plugins/backendconfigpolicy/outlierdetection.go
@@ -16,7 +16,7 @@ func translateOutlierDetection(od *v1alpha1.OutlierDetection) *envoyclusterv3.Ou
 	outlierDetection := &envoyclusterv3.OutlierDetection{}
 
 	if od.Consecutive5xx != nil {
-		outlierDetection.Consecutive_5Xx = &wrapperspb.UInt32Value{Value: *od.Consecutive5xx}
+		outlierDetection.Consecutive_5Xx = &wrapperspb.UInt32Value{Value: uint32(*od.Consecutive5xx)} // nolint:gosec // G115: kubebuilder validation ensures safe for uint32
 	}
 	if od.Interval != nil {
 		outlierDetection.Interval = durationpb.New(od.Interval.Duration)
@@ -25,7 +25,7 @@ func translateOutlierDetection(od *v1alpha1.OutlierDetection) *envoyclusterv3.Ou
 		outlierDetection.BaseEjectionTime = durationpb.New(od.BaseEjectionTime.Duration)
 	}
 	if od.MaxEjectionPercent != nil {
-		outlierDetection.MaxEjectionPercent = &wrapperspb.UInt32Value{Value: *od.MaxEjectionPercent}
+		outlierDetection.MaxEjectionPercent = &wrapperspb.UInt32Value{Value: uint32(*od.MaxEjectionPercent)} // nolint:gosec // G115: kubebuilder validation ensures safe for uint32
 	}
 	return outlierDetection
 }
