+++ internal/kgateway/extensions2/plugins/trafficpolicy/csrf_policy.go
@@ -81,7 +81,7 @@ func constructCSRF(spec v1alpha1.TrafficPolicySpec, out *trafficPolicySpecIr) er
 	// Set filter enabled percentage
 	numerator := uint32(0)
 	if spec.Csrf.PercentageEnabled != nil {
-		numerator = *spec.Csrf.PercentageEnabled
+		numerator = uint32(*spec.Csrf.PercentageEnabled) // nolint:gosec // G115: kubebuilder validation ensures safe for uint32
 	}
 
 	// FilterEnabled is required by the envoy filter and is set to 0 (off) by default
@@ -97,7 +97,7 @@ func constructCSRF(spec v1alpha1.TrafficPolicySpec, out *trafficPolicySpecIr) er
 	if spec.Csrf.PercentageShadowed != nil {
 		csrfPolicy.ShadowEnabled = &envoycorev3.RuntimeFractionalPercent{
 			DefaultValue: &envoy_type_v3.FractionalPercent{
-				Numerator:   *spec.Csrf.PercentageShadowed,
+				Numerator:   uint32(*spec.Csrf.PercentageShadowed), // nolint:gosec // G115: kubebuilder validation ensures safe for uint32
 				Denominator: envoy_type_v3.FractionalPercent_HUNDRED,
 			},
 			RuntimeKey: csrfShadowEnabledKey,
