+++ internal/kgateway/extensions2/plugins/trafficpolicy/extauth_policy.go
@@ -149,7 +149,7 @@ func buildExtAuthPerRouteFilterConfig(
 
 	if spec.WithRequestBody != nil {
 		checkSettings.WithRequestBody = &envoy_ext_authz_v3.BufferSettings{
-			MaxRequestBytes:     spec.WithRequestBody.MaxRequestBytes,
+			MaxRequestBytes:     uint32(spec.WithRequestBody.MaxRequestBytes), // nolint:gosec // G115: kubebuilder validation ensures safe for uint32
 			AllowPartialMessage: spec.WithRequestBody.AllowPartialMessage,
 			PackAsBytes:         spec.WithRequestBody.PackAsBytes,
 		}
